Switches doc building from autosummary to autoapi, which seems to be faster.
Documentation on readthedocs.org is failing to build because the (900s
I think) time limit is being exceeded.  This update should reduce the time
required to build the docs so that readthedocs builds can complete successfully.

Autoapi is a newer sphinx extension that does essentially the same thing as
the older autosummary feature, and may not even require the updated
template we added a little while ago (though this commit leaves it in
and everything seems to work fine).