Fixes (I hope) deployment trigger in Github action; adds "extra" tests back to master.
For our last release (0.9.10), we tried to condition deployment action
(in deploy.yml) on the commit having a tag starting with "v", but this failed
to trigger and we had to upload the action artifact manually.  I think I figured
out the problem: the script was using github.event.ref where it should have been
using github.ref (Note, from https://stackoverflow.com/questions/59349905/
which-properties-does-github-event-in-a-github-workflow-have
that github.ref and github.event.ref differ.  For instance:
github.ref == 'refs/tags/v1.2.5'
github.event.ref == 'v1.2.5').  Hopefully this fixes the issue, and the
next time we tag a release the deployment works automatically.

This commit also re-adds "master" to the branches for which the extras
tests run, as it seems fine and good to re-run all our tests on the
master branch.