Fixes bug in read_circuit_strings & write_circuit_strings duo.
These methods (in readers.py and writers.py) were meant to convert
standard dicionaries and lists potentially containing Circuits to the
"same" JSON-able standard dictionaries and lists with the circuits
replaced by their string reps (and vice versa).  But the implementation
failed to ensure that dictionary keys were strings, and this would
cause non-JSONable errors when using "circuit-str-json" as an auxfile type.

This commit causes dictionaries to be converted into lists marked with
an initial "dict_items" element so that they can be converted back
properly.  This is a bit uglier, but maintains JSONability, which seems
paramount here.