Adds plumbing of circuit_rules in circuit lists (and structures) - and to StandardGSTDesign.
Previously we had just stubbed-out connections for circuit manipulation rules when
creating a StandardGSTDesign.  This commit adds the ability of a CircuitList, and
derived PlaguetteGridCircuitStructure (& its plaquettes), to hold and serialize a
set of circuit manipulation rules (just find/replace pairs for circuits).  The modified
circuits are held as the elements of plaquettes or the actual circuits of a circuit list,
so for most purposes the circuit rules are applied immediately upon creating a circuit
list with rules.  This enables create_lsgst_circuit_lists and StandardGSTDesigns to
correctly utilize circuit rules again (this functionality worked a while ago but was
broken I think when circuit structures were updated).