Adds initial implementation of 'lsvec_mode' for "custom LM" optimizer.
This optimizer setting can take the values "normal" to perform least-
squares optimizations using individual circuit outcome "terms" (as we
have been doing) or a new "percircuit" value which uses the objective
function's per-circuit quantities as the least-squares terms.  This is
useful when, for example, the per-circuit quantities are positive but
the per-circuit-outcome quantities are not, since quantities must be
positive to be compatible with a least-squares optimizer.

Only very light testing has been done so far.