Fixes a bug in convert_strings_to_circuits where dictionary keys could be lists.
Now this function converts a list/tuple (currently ambiguous since JSON
doesn't support distinction) to a *tuple* rather than a list.  This avoids
"list is unhashable" errors when a serialized dictionary had tuple-valued
keys, and I don't think it will cause any problems, as I don't think there are
places where a list is needed over a tuple.  If there are such places, we'll
need to add explicit tags in the serialization.