Adds fallback to __lt__ and __gt__ comparison methods within LabelTupWithArgs.
This allows comparisons to succeed even when there are string/float mismatches.
Such mismatches can occur when some labels have arguments, e.g. ('Gx', 'Q0')
compared with ('Gzr', 3.141592, 'Q0') where the arguments are now stored as
*floats* for efficiency (see prior commit).  This fallback has only been
added to the LabelTupWithArgs class so far, and maybe should be added
to all relevant label classes in the future.