Adds use_op_caching attribute to LocalNoiseModel layer rules.
Adds an expert-functionality flag, use_op_caching, that can be
set to 'False' to disable caching of operations by the layer rules.
This can probably be removed in the future, as this functionality
hasn't thus far been useful for reducing the amount of memory required
for GST with many "gates" (via factories).