Adds additional "nice" serialization routines for operation factories.
Further testing and debugging of serializing (to JSON) models containing
factories revealed several bugs, which this commit fixes.  First, this commit
adds to/from memoized dict methods to the ComposedOpFactory and UnitaryOpFactory
classes.  It also moves the UnitaryGateFunction class (in unitarygatefunction.py)
from the processors subpackage to baseobjs, so that it can be used within
tools/internalgates.py where we need to define top-level classes Gzr and Gczr so
that they can be instantiated by de-serialization routines.