Adds new curly-brace format for allowed gate labels.
Names surrounded by curly braces, e.g. '{myGate}' are now allowed as gate/element names
in pyGSTi.  This allows us to serialize string Label objects that don't meet the usual circuit
syntax rules such as "{auto_global_idle}".

This commit updates the definition of "implicit" or "implied" gates to use curly braces rather
than parentheses (as they did previously), so that implied gate/layer names can be serialized
nicely and not interfere with the parsing of parentheses in circuit strings.

This commit also updates the internal usage of implied gates (the automatic processing of '{idle}'
in explicit models and creation of a '{auto_global_idle}' layer in local noise models) and updates
unit test accordingly.  The ProcessorSpec tutorial is also updated to document the changes.