Fixes a bug in DistributableCOPALayout.fill_jtj when shared memory is unavailable.
An MPI broadcast call was given a transposed numpy array but comm.Bcast just
uses the underlying array data and so this transposition wasn't actually
conveyed to the non-root processors.  This leads to invalid JTJ matrices and
often linear solver failures.  This bug was only in the non-shared-memory part
of the code, and so this only affected cases when shared memory was unavailable
or disabled.