Adds 'pad_1Q' option for implicit_idle_mode arguments for local and cloud-noise models.
This new option creates a layer rule whereby a model's 1Q idle gate is
used to "pad" the idling qubits within a circuit layer.  Entirely empty
"global idle" layers are modeled by composing the 1Q idle on each qubit.
The first single-qubit idle gate (name) given by the creating processor spec
is used for this purpose (idle factories can't be used yet, but could be
added in the future), and the idle gates on different qubits can be different.