Updates & improves serialization of ConfidenceRegionFactory objects.
Adds Hessian projections information and number of gauge/non-gauge
parameters to the information serialized by a ConfidenceRegionFactory.
I think we passed on serializing these at first, thinking the FOGI
work might obviate it quickly.  It seems better to add this fuller
serialization of the object now and just remove it later if we find
it unnecesary.