Fixes bugs within logl and chi2 objective function calls within pygsti.tools
The approximate and actual Hessian-calculation functions in pygsti.tools
(e.g., pygsti.tools.logl_hessian) performed an unnecessary and incorrect
MPI gather operation at their conclusion.  The calls to an objective function
object's .hessian() and .approximate_hessian() methods are different than
its other methods in that these functions gather the (approximate) Hessian
onto the root processor internally, and no additional gathering is needed.
Previously, the code would work in non-MPI cases, since the gather operation
was a trivial pass through, but would fail when using MPI.  This commit fixes
these errors and updates the docstrings to make it clear that the Hessian
matrix is only returned on the root processor.