Adds more options for "builtin" ideal prep and POVM types (for processor specs).
Adds recognition of strings that specify standard z-basis preps and
POVMs.  This is an initial pass, and adds recognition of strings like
"rho3", "rho011", "011", and "E3".  You can also pass numpy arrays for
the pure state vector or effect vector.  It seems useful to be able to
specify a multiple "pure" effects that get added into one POVM effect,
but this isn't implemented yet.

This commit also adds unit tests for state construction to ensure
conventions hold up as we expect them to.  For example, a 3-qubit
state "rho4" is the same as "rho_100" and we list qubits in order
left-to-right (the 0-th qubit is 1 in "rho_100") and, perhaps
non-intuitively we increment the highest-indexed qubit *first*.
For example, state "1" in a 3-qubit system is "rho_001" which has
qubit0=0, qubit1=0, qubit2=1.