Adds support for 'builtin' instrument members and expands POVM spec format.
Adds ability to specify instrument members within QubitProcessorSpec objects
as pairs of effect&prep specifiers, which can now be of the form,
e.g., ('00', '00') or ('E0','rho0') or (<ndarray>, <ndarray>).
(Previously only "rho0" was recognized as a temporary placeholder.)

Secondly, this commit expands the specification of POVM effects within
processor specs to be a *list* of state-projection effects to create
an effect that is the *sum* of those projectors, similar to the way
instruments are specified.  This will allow for more general types
of ideal POVMs to be specified.

Finally, this commit adds several unit tests that check that SPAM
and instrument elments can be successfully added to processor spec
objects, and that 1) these objects can be serialized and 2) they can
be used to create explicit, local, and cloudnoise models.  The tests
aren't incredibly extensive, but they should verify the correct operation
of the basic usage of SPAM and instrument elements with processor specs.