Fixes factory-optimization bugs causing errors in probability computations.
The main issue this commit resolves is adding a cast from a complex->real
numpy array when appropriate, as a part of an expanded call to basistools.change_basis
within UnitaryOpFactory.create_object.  This was significant because the dense
rep blindly takes a numpy array and treats its data as real, so the element size
was wrong and this thereby messed up all the dense reps and therefore the probability
computations.

This commit also fixes a syntax error in DenseUnitaryOperator.quick_init.