Adds 'orthogonalize' argument to nice_nullspace and auto-selects FOGI norm order.
In order to make FOGI analysis capable of replicating the past SPAM-free
analysis of UNSW data, several tweaks were (and maybe more are) needed.

First, the 'nice_nullspace' function would sometimes construct a non-orthogonal
nullspace when working on the expanded H+S space (rather than H or S separately),
and now an additional 'orthogonalize' argument, which default to False, adds a Gram-Schmidt
orthogonalization as a part of the construction.  This commit also fixes a minor bug
in the nice_nullspace normalization, making the normalization choice more robust.

Secondly, since the H and S FOGI analyses are no longer separate (like they
were before we added support for SPAM elements), choice of the "norm_order"
(1 for S-like quantities, 2 for H-like) needed to be chosen on the fly based
on the type of FOGI direction being dealt with.  The construct_fogi_quantities
method now accepts "auto" as a value for its norm_order argument, which selects
the appropriate norm if and when a FOGI vector is all-S-type or all-H-type.  If
a vector is mixed H+S then norm_order=2 is the default.