Fixes serialization of DensePureState, allowing complex numbers and to_dense('Hilbert').
Similar to an earlier fix for DenseUnitaryOp I think.  Previously the
state vector array was assume to be real, which of course it isn't always.
This commit also enables a pure state represented by a "superket" (e.g.
within the "densitymx" evotype) to be given as a dense *pure state* as
is needed to serialize it.