Adds methods to ExplicitModel and FOGI store classes to process/aggregate FOGI errors.
The added functions, some of which are H+S specific, bring into pyGSTi some of
the analysis we found useful in specific cases (e.g. UNSW data).  These additions
will likely need to be tweaked but provide a foundation for the sort of FOGI-error
aggregation and presentation we'd like to be a part of pyGSTi (and in reports in
the future).