Updates FullUnitaryOp and .size -> .hilbert_schmidt_size so statevec evotype works.
Fixes DenseUnitaryOp.num_params, which returned "self.size * 2", which translated
to self.dim**2 * 2 which was incorrect because self.dim is always the
Hilbert-Schmidt dimension (it's the op's .state_space.dim not .udim).  To make things
clearer, this commit updates the .size attribute (which could sometimes be confused
with a numpy's .size attribute) of model member objects to .hilbert_schmidt_size,
and specifically fixes DenseUnitaryOp.num_params to be self._ptr.size * 2.

*NOTE* A potentially lurking problem is that LinearOperator.dim gives the Hilbert-Schmidt
dimension and there is no LinearOperator.udim.  It's possible (I forget) that
one or both of the Matrix and Map forward sims uses .dim in a way that assumes
it will be the *unitary* dimension rather than the Hilbert Schmidt dimension.  Maybe
we want a .udim or rename .dim -> .hilbert_schmidt_dim?

There was a problem using unitary operations with evotype='statevec'