Begins work adding better qudit support to pyGSTi.
This commit begins by adding the QuditSpace state space and
the QuditProcessorSpec classes, paving the way to upates to model
construction for a system of qudits rather than just qubits.  The
QuditSpace is still less general than an ExplicitStateSpace in that
it's still just a single tensor product (no direct sum) of qudit
spaces.