Partially adds ability to use qudits (rather than just qubits) in model constructions.
Adds QuditSpace and QuditProcessorSpec classes, and updates model construction
methods to be compatible with qubit processor specs.  Implementation is "partial"
in that we still need to update hardcoded references to the Pauli product ("pp") basis
throughout model construction (including noise construction) as this is likely
incompatible with a space of qudits.

Updates reference files (needed because QubitSpace needs to have
a .qudit_labels attribute) so that unit & system tests pass.