Revamps LindbladErrorgen and adds 'C' and 'A' type elementary error generators.
Replaces H and S/other error generator creation functions (some of which
generated dual, some primal "vectors", but naming was confusing) with new
create_elementary_errorgen_dual, create_elementary_errorgen, and
create_lindbladian_term_errorgen functions.  The first two clearly construct
the dual/primal H, S, C, and A-type "elementary" error generators, which are
distinct from the "lindblad term" error generators (corresponding to a
given H or "other"/"non-ham" term of a Lindbladian expansion) created by
the final function (this function constructs *primal*) vectors, and there's
no need ot construct a dual basis to this since we never need to project onto
these - we just convert the projections onto elementary error generators to these.
This commit removes the weird "std_scale_factor" and scaling within the elementary
errorgen computation - now that we have separate and well defined primal & dual
functions this shouldn't be an issue.  I did need to compute the normalization factor
automatically within create_elementary_errorgen_dual as this varies depending on the
normalization of the basis (e.g. normalized Paulis vs traditional Paulis).

Secondly, many functions in lindbladtools.py and optools.py related to converting
lindblad-term coefficients to/from dictionaries and vectors of parameters have been
replaced by the new LindbladCoefficientBlock class (in lindbladcoefficients.py).  The
LindbladErrorgen class now holds a list of these blocks instead of calling the tools
functions, and has been massively rewritten by this commit.  The result is a changed
__init__ signature and increased flexibility, as the version of LindbladErrorgen
prior to this commit essentially hardcoded the use of two blocks - a "ham" and
an "other_diagonal"/"other" block.  Now a LindbladErrorgen object uses its arbitrary
list of LindbladCoefficientBlocks to generate its parameters and perform derivative
calculations, etc.

Lindblad-term error generator coefficients are still used (and LindbladErrorGen
is based on blocks of *Lindblad-term* coefficients) rather than just using elementary
error generators (H, S, C, & A types) everywhere because the former allow us to
enforce the CP constraint straightforwardly.  When this constraint is not needed,
we could just paramterized H, S, C, A coefficients directly, and in the future we
could make a new/derived LindbladCoefficientBlock object for this.  Still, the user
can interact with the coefficient blocks and LindbladErrorgen completely in terms of
elementary error generators, and so the Lindblad-term generators can be seen as an
under-the-hood detail used to enforce CP.

At this point, a number of related changes need to be made to related objects and
functions, and unit tests will need to be adjusted.  This commit will break normal
functioning and is a work in progress.