Adds non-normalized Pauli basis ('PP') and fixes bugs & tests related to LindbladErrorgen.
Adding a non-normalized Pauli basis is especially useful as an
elementary errorgen basis, and is used by default in stochastic and
depolarization ops to realize the expected "correct" error rate scale.

This commit makes several minor but important tweaks to the new
elementary errorgenerator construction routines, LindbladCoefficientBlock,
and LindbladErrorgen so that they seem to be working correctly.  After
updates to the unit tests, all unit tests pass now.  System tests still
need to be updated and haven't been checked yet.