Fixes several bugs related to recent LindbladErrorgen updates.
The main change of this commit is additional logic around the Cholesky
decomposition that separately pulls out exactly-zero "other" parameters
and doesn't include them in the call to scipy's Cholesky decomposition
routine.  This replaces the HACK of pushing zero eigenvalues of the
"other" Lindbladian coefficient matrix positive (by 1e-12 or 1e-16),
and so should make the creation and optimization of LindbladErrorgen
objects more sensible and robust.

A side effect, however, was that this uncovered a bug in the path-integral
forward simulations (needed to add LOGSMALL to termforwardsim_calc_* modules)
and adds a need to seed path integral optimization slightly away from the
all-zero parameter vector of the target model (for reasons that aren't 100%
clear yet, but have to do with the optimizer not constructing a good derivative
at the start of the optimization otherwise).

Another update in this commit is the replacement of 'pp' with 'PP' as the
default elementary/lindblad errorgen basis when converting/constructing
error generators in various places.  This causes the coefficients to match
those previously used in pyGSTi up to factors of 2 for the H-type generators.

This commit also includes minor fixes and updates discovered by running the
system tests (which themselves still need to be updated).  Most of these
involved replacing old (now commented out) error generator routines with
use of new routines or LindbladCoefficientBlock object usage.