Updates idle tomography and reports for new LindbladErrorgen; updates system tests.
Updates idle tomography and report functions to use updated functions
(& LindbladCoefficientBlock objects) for dealing with elementary error
generators.  This work includes updating the table of elementary errorgen
coefficients within the report so that it now shows C and A rates (and no
more "affine"), and replaces "Power" with "Fraction" to be more clear.
These updates are more quick & dirty than well thought out, as we plan to
revamp the reports in the near future.

Secondly, this fixes a forgotten-matrix-transposition bug in the
permutation logic surrounding the Cholesky decomposition introduced in
a recent commit.

This commit also updates the system tests as needed, and 1Q calc method
reference files (that still rely on pickled models!).  At this point, all
the unit tests and all the system tests except interpygate tests that were
already broken are passing on Erik's machine.