Updates many instances of dtype=int to dtype=np.int64 to fix Windows issues.
Using 'int' as a dtype is ambiguous and Windows sometimes/always uses a
32 bit integer whereas OSX/Linux use 64 bit integers, causes issues on
Windows particularly when Cython code is called.  This commit makes lots
of replacements that should get rid of most/all usage of 'int' as a numpy
dtype, and as extra measure it adds dtype=np.int64 args to the
ascontiguousarray calls within the stabilizer evotype's clifford oprep.