Fixes bugs caught by system tests after adding 'pad_1Q' implicit idle mode.
In particular, fixes a bug in term-calcs when 'implicit_idle_mode = None'
(a side effect of the actual updates, this bug was caught because the
default changed from "add_global" to "none" for a cloudnoise model).

Also adds a special case for idle tomography and cloud circuit construction
when using '{idle}' gates:
Idle tomography and cloud circuit construction queru the processor spec for
the name of the idle gate and use this to construct circuits.  BUT, if the
model is a cloud noise model that adds a global idle to all layers, then the
*layer* op returned for the ['{idle}'] circuit will have two applications of
the idle gate (one as the added "global idle") when we only want application.
This commit adds a hack that detects and deals with this case.