Updates CompleteElementaryErrorgenBasis to and FOGI construction to work with H,S,C,A generators.
After recent merge of updated FOGI calculations, this commit updates
a central piece of the FOGI framework - the CompleteElementaryErrorgenBasis -
to use the *new* H/S/C/A elementary generators instead of the old (and now
largely absent) "Lindblad term" generators.  This commit also updates the
`coefficient_labels` and `coefficient_array` methods of LindbladErrorgen so
that they give *elementary* errorgen labels as opposed to Lindblad term labels,
which seems best as this is now consistent with the `coefficients` method.

This change, however, required updates to the derivative-of-coefficients code in
LindbladErrorgen and LindbladCoefficientBlock, and while this code runs it may
contain errors (hasn't been tested).  In fact, trying to use FOGI reparameterization
dies with a singular matrix linear algebra error, so it's likely soemthing
isn't quite right with this derivative computation (FUTURE work).