Adds HACK in errorgen-projections report calcs for using new H/S/C/A errogens.
After updating the types of elemetary error generators and how they are
constructed (namely their scaling), the error generator coefficients that
we used to report (as scaled versions of elementary errorgens based on
normalized Pauli matrices) are now equal the coefficients of the elementary
errorgens based on the normal, not-normalized Pauli matrices.  This commit
adds a hack that uses the un-normalized Paulis ('PP' basis) when projecting
the error generators of models that have a normalized Pauli ('pp') basis or
a product of 'pp' bases.

In the FUTURE the errorgens_and_projections routine in reportables should
either be passed the projection basis directly or compute it in a more
sophisticated way, e.g., looking at the error generators of the model's
operations, to remove this HACK.

NOTE: the elementary errorgen coefficients for models with non-Pauli bases
will *change* in the reports, as these bases are passed directly through and
used to construct elementary errorgens which now are constructed with a
*different* scaling.