Updates EigenvalueParamDenseOp so that n x n degenerate blocks are parameterized.
Previously when constructing a EigenvalueParamDenseOp parameters would
be added for degenerate 2x2 blocks by setting include_off_diags_in_degen_2_blocks=True.
This, however, was insufficient for correctly performing per-germ-per-L FPR,
which needs, e.g., to know about the 3x3 (and larger) degenerate blocks of
TP identity/idle gates on 1 or more qubits.  This commit renames the argument to
include_off_diags_in_degen_blocks and updates the code so that it correctly
adds all the parameters needed to parameterize the commutant of the gate matrix
it's given (all the off-diagonal elements of a degenerate block).