Updates the per-germ-per-L FPR algorithm to test eigenvalue magnitude and not just rank.
Previously a set of fiducial pairs was considered "good" as soon as its rank
was equal to the number of (EigenvalueParamDenseOp) parameters that need to
be amplified.  This may have worked ok, but the new implementation should be
significantly more robust: it checks the lowest eigenvalue corresponding to
an amplified parameter and selects 1) only fiducial pair lists where this
lowest-eigenvalue is above some threshold and where the condition number is
less than a threshold and 2) the best among several (minimally 10 or the number of
random instances / 2) pair lists.

Furthermore, the per-germ-per-L FPR algorithm sets the lowest-eigenvalue
threshold for L > 1 cases on the expected increase in sensitivity from
lower-L cases already performed.  This gives further certainty that the
fiducial pairs being selected are as sensitive to the amplifiable parameters
as they should be.

There are currently a number of HARDCODEed parameters marked in the
FPR algorithm that should work fine in most cases, but may need to be
plumbed up to being adjustable by the user in the FUTURE.