Add Gu3 to Circuit.convert_to_openqasm.
This change requires an overhaul of the std_gatenames_to_argmap,
since Gu3 for the x-sx-rz is not a single gate.
The new format uses functions to generate the entire list of qasm
instead of a single gate.
This change also removes the need for the args to be strings
for argmap to work properly. Args are cast to str as needed.