Updates default gauge optimization settings and improves conversion routines.
Adds a new argument, `convert_model_to` to gaugeopt_to_target.  This allows
the user to specify that before being gauge-optimized a model should be converted
to a given operation/SPAM object type *and* have its default gauge group reset
appropriately.  The default 'stdgaugeopt' gauge optimization suite is updated
to convert models to "full TP" before gauge optimizing them (though the gauge
groups are still chosen according to the initial model's gauge group).

This substantial update in behavior is motivated by two recent discoveries:

1) Performing Linbladian-constrained ("CPTP"-constrained) gauge optimization,
even over the unitary group, can and does result in a model being "non-Lindblad".
This causes the gauge optimization to fail (or stop) unless the tranformation
to Lindblad-form is truncated (which it was with a tolerance of 1e-4).

2) Allowing such a large tolerance (1e-4) caused the gauge optmization of a
"CPTP" model to slightly increase the 2DeltaLogL of the model & data.  This
was thought to be no big deal, but investigations of the logL's Hessian
(used for computing error bars) showed otherwise -- the specturm of the Hessian
is significantly affected by not being at the actual (to the optimizer tolerance)
peak/ridge of the logL, especially when using the "CPTP" (Lindbladian)
parameterization since the gauge/non-gauge gap in its spectrub is less robust
that, say, a TP-constrained model.

In addition to changing the defaults as mentioned above, the Lindblad truncation
tolerance is dropped from 1e-4 to 1e-12 to lessen the ill-effects of gauge optimizing
a Lindblad-constrained model.  This commit also adds a new gaugeopt suite called
"stdgaugeopt-noconversion" which does *not* convert the model to "full TP" before
gauge optimizing.  This will partially replicate old behavior, but not exactly as
the drop in tolerance will make using this suite more error/stopping prone.

The second update contained in this commit regards the conversion routines
found in the various modelmembers/*/__init__.py files.  These files used to
contain separate code for converting a) a dense matrix/vector to a modelmember
object and b) an existing modelmember object to one of another type.  This resulted
in significant code duplication, and this commit updates the covert(...) routines
(which perform action b) to use the matrix/vector creation routines (action a)
except in special circumstances.  This update also makes both sets of routines
more flexible and robust.  The function signature for the convert(...) functions
is updated to include new `ideal_*` and `flatten_structure` arguments that
access some of this added flexibility.

Finally, ExplictModel has a new method, `convert_members_inplace` that is
more flexible than, and now utilized by, the set_all_parameterizations
method.  In the future, we may be able to move this method upward into
OpModel and implement it within implicit models as well, which would be
nice.

Unit and system tests still pass, at least on Erik's machine.