Fixes bugs in the way the Hessian matrix is projected within a ConfidenceRegionFactory.
Prevously we would project a Hessian matrix, H, by dotting it on
both sides with the linear projector onto a selected non-gauge space,
H -> Png @ H @ Png.  This was incorrect, as the Hessian behaves differently
(and is more fragile) under coordinate transformations.  If B gives a
coordinate transformation v -> B @ v, then a Hessian transforms as
H -> inv(B).T @ H @ inv(B).  This commit fixes this issue and adds a
check that the Hessian on the gauge space and between the gauge and
non-gauge spaces is small (~ magnitude of the gradient, as expected)
making the Hessian near block-diagonal in the (nongauge, gauge) basis.
A warning is issued when this is not the case.

This commit also adds a temporary check that we understand why the Hessian
on the gauge space isn't exactly zero (and is ~ size of the gradient).  It
computes the gauge-orbit curvature tensor "Q" (a tensor of 3rd derivatives of
logL) which gets dotted into the gradient to produce an expected correction
to the Hessian on the gauge space.  Testing in several 1Q cases confirms that
this correction accounts for the non-zero Hessian on the gauge space (yay!).
This experimental code is marked with "CORRECTION" tags, and will be removed
in the next commit, as it just confirms understanding and isn't needed in the
analysis.  If we want to test this later on, this commit should be pulled
or the difference between this commit and the next should be examined.

Note that this commit updates the Hessian projection routines in ConfidenceRegionFactory,
which are used to compute error bars, and so computed error bars should
be expected to change after this commit.