Fixes a bug in bootstrap unit testing.
We were using a LGST model instead of a unitary-gate-containing
target model in the bootstrap tests.  This commit simply replaces
that model with the true target model.  If there's a reason we need
to supply non-target models to the bootstrap routines as their
`input_model` argument, we should also allow `target_model` to be
specified (currently you can't specfiy both), as a needed processorspec
can only be build from a unitary-containing target model.