Adds Basis.is_equivalent to allow for sparse-agnostic Basis comparisons.
When running multi-qubit GST, particularly with path integral calculations
we can have a mix of sparse and non-sparse bases based on not-easily-controllable
context such as what the available rep types are within an evotype.  This makes
it useful to be able to compare Basis objects in a way that ignores whether their
`sparse` flags match.  The new Basis.is_equivalent method implements the existing
__eq__ methods and can be called separately with `sparseness_must_match=False`
to compare to Basis objects in a sparse-agnostic way.  This functionality is
used within ComposedErrorgen construction, where an assertion was failing because
some of the composed members had a sparse vs non-sparse version of the same basis.