Updates FOGI analysis to fix sign convention in QR decomposition.
This causes more stability in the signs of created FOGI directions.
This is useful, for example, when construting FOGI quantities for
two similar but not quite identical models, where you'd expect the
FOGI quantities to be the same but before this commit they might
have FOGI directions that differ in a sign flip (making it annoying
to compare the FOGI coefficients of the models).