Adds warning when calling LindbladErrorgen.__init__ with a dictionary.
Directs the user to use the from_elementary_errorgens method, since
__init__ was recently updated to take a list of LindbladCoefficientBlock
objects instead of a list of elementary errorgens.

This commit also moves the state space cast within
LindbladErrorgen.from_elementary_errogens a few lines higher so
that it actually works (so you can, e.g., supply an integer as the
state space).