Updates use of operation factories so that zero-argument factories are allowed.
A minor update to the way op factories are used to construct an operation from
a given label.  Previously, only if a Label had 1 or more arguments did we look
for a factory object with key == the label without its arguments (typically just
a name and sslbls).  Now this is performed even when the label has no arguments,
allowing the use of zero-argument factories that are also specific to a set of
state space labels.  (E.g. previously if model.factories[('Gfact',0)] was a factory
that expected no arguments it wouldn't be usable.)

This update is slightly controversial, since there really isn't much point to having
factories without arguments that are just for a single set of sslbls (contrast this
with factories that *do* depend on sslbls, like an EmbeddingOpFactory).  Allowing for
this, however, doesn't cause any harm to other functionality and may be desirable just
to maintain a common coding paradigm when using pygsti.