Updates vbplot.py to avoid unnecessary imports of matplotlib and seaborn.
This fixes an issue brought up by a user in which pygsti, when imported,
changes the default matplotlib parameters (and it obviously shouldn't).
This was happening because seaborn does this when it's imported and
vbplot.py attempted to import seaborn when it was loaded, regardless of
whether it was ever needed.  This commit pushes these imports back into
the functions that actually use them, so this unexpected behavior will
only occur when you're creating volumetric benchmarking plots - not after
just importing pygsti.