Removes HasProcessorSpec.create_target_model and improves target model handling.
When shifting from target models to processor specifications, many places where
a target model was still needed were replaced with `pspec.create_target_model()`.
This was fine as a stopgap, but within pyGSTi a model is typically created *from*
a processor spec and not the other way around.  This is because a Model contains
additional information such as the parameterization, basis, state space, and
functionality for computing circuit outcome probabilities.  Thus, a processor spec
object should *not* be able to create a Model.  In many places where a target model
was used only a processor spec was necessary, and there the previous replacements
are legitimate.  In this commit, the other places - where a target model is actually
needed - are addressed.  The stopgap pspec.create_target_model() function has been
removed (raises a NotImplementedError) and several functions/objects now require
a target model be specified instead of just a processor spec.  These include:

- driver functions run_stdpractice_gst and run_linear_gst now take a target
  Model (or path to one) instead of a processor spec, essentially reverting
  back to the old function signature.  If you pass a processor spec as before,
  a warning is issued and the pspec is used to build an explicit model, so
  that backward compatibility is preserved.

- the StandardGST protocol object now takes and stores a target_model which
  serves as the base for creating various parameterizations.  Even though
  the parameterization of this model doesn't matter, its basis (and perhaps
  other attributes) does.  Leaving target_model=None will cause the Protocol
  to construct an explicit model from the edesign processor spec, so a target
  model is not stricly necessary if explicit models are being used.

- in ModelTest objects, a target model is not constructed from the edesign
  anymore.  A target model is included in the results only if one was given
  to the protocol when it was constructed.