Adds 'basis' argument to TPState constructor, fixing a bug in TP test.
Previously the TP check consisted of testing that the first element
was dim^-(1/4) (so 0.707.. for 1 qubit), where dim == length of the
superket vector.  This doesn't work, however, when using a kite basis
where the basis dimension (== length of superket vector) is smaller than
the full density matrix Hilbert space.  Instead, the test should use
(I think) basis.elsize, which is now does and requires the additional
'basis' argument that wasn't needed before.