Updates PyPI deploy workflow actions after failed auto-deployment of v0.9.10.1.
After creating and pusing the v0.9.10.1 tag, the existing deploy.yml didn't
trigger, I guess because the github.ref didn't match /ref/tags/v*?  After the
failure, I had to manually run the testPyPI deployment and upload those files
manually to PyPI to create release, which was annoying.  To try to avoid this
in the future, this commit renames deploy.yml -> autodeploy.yml and changes
(again) the way the triggering is done, so hopefully for the next release it
triggers properly.  In case it doesn't, this commit also adds manualdeploy.yml,
which is identical to autodeploy.yml but is configured to only be run manually.
So if the auto deployment doesn't trigger, getting the release to build and
upload should be as easy as running the manual action.